requirements.txt
# =============================
# Core AI / ML
# =============================
torch==2.2.2+cpu
--extra-index-url https://download.pytorch.org/whl/cpu

transformers>=4.40.0
numpy>=1.26.0
scikit-learn>=1.4.0

# =============================
# Media Processing
# =============================
pillow>=10.0.0
opencv-python-headless>=4.9.0
speechrecognition>=3.10.0

# =============================
# API
# =============================
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
slowapi>=0.1.9
pydantic>=2.0.0

# =============================
# Optional (lighter versions)
# =============================
librosa>=0.10.0

# =============================
# Utilities
# =============================
python-dotenv>=1.0.0
requests>=2.31.0

# =============================
# AI News Pipeline
# =============================
feedparser>=6.0.0
apscheduler>=3.10.0
google-api-python-client>=2.99.0
google-auth-oauthlib>=1.0.0
